# docs/api/openapi.yaml
openapi: 3.1.1
info:
  title: Sick-Fansubs HTTP API
  version: "1"
  description: >-
    Accepted Sick-Fansubs wire contracts. Product APIs use /api/v1; operational
    health endpoints are intentionally outside product API versioning.

    Cross-cutting: every endpoint whose chain carries the forced-change gate
    (the authenticated reads chain and the authenticated chain) additionally
    answers 403 with problem type /problems/auth/password-change-required for
    sessions whose account carries the forced-change flag. The /api/v1/auth/
    subtree never carries the gate, so its paths stay reachable while flagged —
    including PUT /api/v1/auth/password, the sign-out pair, the session
    bootstrap, POST /api/v1/auth/reset-password, and
    POST /api/v1/auth/resend-verification.

    Cross-cutting: every documented unsafe operation (POST, PUT, PATCH,
    DELETE) requires the trusted origin and, when a session is present, the
    session-bound X-CSRF-Token header matching the csrfToken returned by the
    session bootstrap — a missing, malformed, or mismatched token is a 403.
    Safe methods never require it, and an unauthenticated unsafe request
    (for example sign-in, register, forgot-password, or reset-password) is
    protected by the trusted origin alone. An expired session has the
    clearing sf_session cookie written before its handler runs, and the
    operations whose handler clears the cookie — sign-out, sign-out-all,
...
    their own response.

...
servers:
...
paths:
...
  /api/v1/media:
    post:
      operationId: uploadMedia
      summary: Upload and process a media image (admin+)
      description: >-
        The upload-first flow. Multipart/form-data with a single "file" part —
        extra fields are rejected. The whole body is bounded (10 MB file +
        framing); the declared Content-Type is ignored (JPEG/PNG/WebP input
        is detected from magic bytes). The file is processed to the 1280×720
        JPEG-85/PNG target and stored under a fresh random 32-hex id.
        Requires an admin or super-admin session, the trusted origin, and a
        valid X-CSRF-Token. 201 returns the storage-relative path (the value
        the content forms submit as thumbnailPath) plus the absolute served
        URL; Location is the served URL path. A source over 10 MB, over the
        2^24 total-pixel bound, or over the 8192 per-axis bound is a 413
        with violation file:tooLarge; undecodable bytes are a 422 with
        file:invalidFormat; a missing file part is a 422 with file:required.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              additionalProperties: false
              properties:
                file:
                  type: string
                  format: binary
                  description: The image file (at most 10 MB).
      responses:
        "201":
          description: Image processed and stored.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestID"
            Cache-Control:
              $ref: "#/components/headers/NoStoreCacheControl"
            Location:
              description: The served URL path of the new image.
              schema:
                type: string
                pattern: "^/media/images/[0-9a-f]{32}\\.(jpg|png)$"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MediaUploadResult"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/InvalidCredentials"
        "403":
          $ref: "#/components/responses/Forbidden"
        "413":
          $ref: "#/components/responses/MediaTooLarge"
        "422":
          $ref: "#/components/responses/Validation"
        "500":
          $ref: "#/components/responses/InternalError"
  # ── Media explicit delete ──
  /api/v1/media/{file}:
    parameters:
      - name: file
        in: path
        required: true
        description: >-
          The served media segment — the last path segment of the upload
          response's url field (<32hex>.<ext>), never hand-built. The
          pipeline writes jpg/png only; .jpeg is parser-tolerated for
          hand-placed legacy files and never matches a written file.
        schema:
          type: string
          pattern: "^[0-9a-f]{32}\\.(jpg|jpeg|png)$"
    delete:
      operationId: deleteMedia
      summary: Delete an uploaded media image (admin+)
      description: >-
        The explicit-deletion terminal state of the upload lifecycle.
        Requires an admin or super-admin session, the
        trusted origin, and a valid X-CSRF-Token. A file referenced by any
        active content row or user avatar is never removed (409
        /problems/conflict); an
        unreferenced file is removed (204). A malformed segment is 400, a
        well-formed unknown file 404. No audit event — deletion is logged,
        not audited (the audited-event list is exhaustive; the upload rules apply unchanged).
      responses:
        "204":
          description: File removed.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestID"
            Cache-Control:
              $ref: "#/components/headers/NoStoreCacheControl"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/InvalidCredentials"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: >-
            The file is referenced by at least one active content row or
            user avatar — it
            is never removed. Problem type /problems/conflict (the shared
            Conflict component's "duplicate race" wording does not apply
            to this endpoint; the operation description is authoritative).
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestID"
            Cache-Control:
              $ref: "#/components/headers/NoStoreCacheControl"
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  headers:
    RequestID:
      description: Server-generated request correlation identifier.
      schema:
        type: string
    NoStoreCacheControl:
      description: API responses default to no-store.
      schema:
        type: string
        const: no-store
...
  schemas:
...
    Problem:
      type: object
      description: >-
        RFC 9457 Problem Details with only accepted safe extensions. Endpoint
        contracts define their stable problem types and violation codes.
      additionalProperties: false
      required:
        - type
        - title
        - status
        - requestId
      properties:
        type:
          type: string
          format: uri-reference
          description: Stable relative problem type identifier.
          example: /problems/validation
        title:
          type: string
          description: Safe developer-oriented summary; not localized UI copy.
        status:
          type: integer
          minimum: 100
          maximum: 599
          description: HTTP status code for this occurrence.
        requestId:
          type: string
          description: Server-generated request correlation identifier.
        violations:
          type: array
          description: Deterministically ordered field violations when applicable.
          items:
            type: object
            additionalProperties: false
            required:
              - field
              - code
            properties:
              field:
                type: string
                description: Accepted public request-field identifier.
              code:
                type: string
                description: Stable machine-readable violation identifier.
...
    MediaUploadResult:
      type: object
      description: >-
        The 201 body of POST /api/v1/media. path is the
        storage-relative value the content create/update DTO accepts as
        thumbnailPath; url is the absolute served URL.
      additionalProperties: false
      required:
        - path
        - url
      properties:
        path:
          type: string
          pattern: "^media/images/[0-9a-f]{2}/[0-9a-f]{32}\\.(jpg|png)$"
          description: Storage-relative media path (never hand-built by clients).
        url:
          type: string
          format: uri
          description: Absolute served URL of the processed image.
...
  responses:
    BadRequest:
      description: Malformed request (400) — problem type /problems/bad-request.
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestID"
        Cache-Control:
          $ref: "#/components/headers/NoStoreCacheControl"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    InvalidCredentials:
      description: Generic 401 — problem type /problems/auth/invalid-credentials.
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestID"
        Cache-Control:
          $ref: "#/components/headers/NoStoreCacheControl"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
...
    Forbidden:
      description: >-
        403 — problem type /problems/forbidden (CSRF/origin failure or a role
        floor) or /problems/auth/password-change-required (the forced-change
        gate — see PasswordChangeRequired).
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestID"
        Cache-Control:
          $ref: "#/components/headers/NoStoreCacheControl"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
...
    MediaTooLarge:
      description: >-
        413 upload over the 10 MB file bound or the multipart framing bound
        (the file bound plus 1 MiB), or over the 2^24 total-pixel or
        8192-per-axis processing limits — problem type /problems/validation
        with violations [{file, tooLarge}].
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestID"
        Cache-Control:
          $ref: "#/components/headers/NoStoreCacheControl"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
...
    NotFound:
      description: >-
        404 — problem type /problems/not-found: an unknown path under /api/v1
        (namespace catch-all), or an absent target on a resource operation
        (the operation's description states what its 404 discloses).
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestID"
        Cache-Control:
          $ref: "#/components/headers/NoStoreCacheControl"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    Validation:
      description: 422 with field violations — problem type /problems/validation.
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestID"
        Cache-Control:
          $ref: "#/components/headers/NoStoreCacheControl"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
...
    InternalError:
      description: 500 generic unexpected failure — problem type /problems/internal-error.
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestID"
        Cache-Control:
          $ref: "#/components/headers/NoStoreCacheControl"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"